move folders up one level
