+++ src/app/app.module.ts
@@ -10,16 +10,16 @@ import { AppRoutingModule } from './app-routing.module';
 import { TranslateModule } from '@ngx-translate/core';
 import { VirtualScrollerModule } from 'ngx-virtual-scroller';
 
-import { AlphabetPrefixService } from './components/pipes/alphabet-prefix.service';
+import { AlphabetPrefixService } from './pipes/alphabet-prefix.service';
 import { AutoTagsSaveService } from './components/tags-auto/tags-save.service';
 import { AutoTagsService } from './components/tags-auto/autotags.service';
 import { ElectronService } from './providers/electron.service';
 import { HomeComponent } from './components/home.component';
 import { ManualTagsService } from './components/tags-manual/manual-tags.service';
-import { ResolutionFilterService } from './components/pipes/resolution-filter.service';
-import { ShowLimitService } from './components/pipes/show-limit.service';
-import { StarFilterService } from './components/pipes/star-filter.service';
-import { WordFrequencyService } from './components/pipes/word-frequency.service';
+import { ResolutionFilterService } from './pipes/resolution-filter.service';
+import { ShowLimitService } from './pipes/show-limit.service';
+import { StarFilterService } from './pipes/star-filter.service';
+import { WordFrequencyService } from './pipes/word-frequency.service';
 
 import { AddTagComponent } from './components/tags-manual/add-tag.component';
 import { AppComponent } from './app.component';
@@ -33,7 +33,7 @@ import { IconComponent } from './components/icon/icon.component';
 import { MetaComponent } from './components/meta/meta.component';
 import { PreviewComponent } from './components/views/thumbnail/preview.component';
 import { SheetComponent } from './components/sheet/sheet.component';
-import { SimilarityService } from './components/pipes/similarity.service';
+import { SimilarityService } from './pipes/similarity.service';
 import { SliderFilterComponent } from './components/slider-filter/slider-filter.component';
 import { StatisticsComponent } from './components/statistics/statistics.component';
 import { SvgDefinitionsComponent } from './components/icon/svg-definitions.component';
@@ -41,26 +41,26 @@ import { TagsComponent } from './components/tags-auto/tags.component';
 import { TopComponent } from './components/top/top.component';
 import { ViewTagsComponent } from './components/tags-manual/view-tags.component';
 
-import { AlphabetPrefixPipe } from './components/pipes/alphabet-prefix.pipe';
-import { CountPipe } from './components/pipes/count.pipe';
-import { FileSearchPipe } from './components/pipes/file-search.pipe';
-import { FileSizePipe } from './components/pipes/file-size.pipe';
-import { FolderArrowsPipe } from './components/pipes/folder-arrows.pipe';
-import { FolderViewPipe } from './components/pipes/folder-view.pipe';
-import { LengthFilterPipe } from './components/pipes/length-filter.pipe';
-import { LengthPipe } from './components/pipes/length.pipe';
-import { MagicSearchPipe } from './components/pipes/magic-search.pipe';
-import { ManualTagSortPipe } from './components/pipes/manual-tags-sort.pipe';
-import { ResolutionFilterPipe } from './components/pipes/resolution-filter.pipe';
-import { ReturnZeroPipe } from './components/pipes/returnZero.pipe';
-import { SimilarityPipe } from './components/pipes/similarity.pipe';
-import { SortingPipe } from './components/pipes/sorting.pipe';
-import { StarFilterPipe } from './components/pipes/star-filter.pipe';
+import { AlphabetPrefixPipe } from './pipes/alphabet-prefix.pipe';
+import { CountPipe } from './pipes/count.pipe';
+import { FileSearchPipe } from './pipes/file-search.pipe';
+import { FileSizePipe } from './pipes/file-size.pipe';
+import { FolderArrowsPipe } from './pipes/folder-arrows.pipe';
+import { FolderViewPipe } from './pipes/folder-view.pipe';
+import { LengthFilterPipe } from './pipes/length-filter.pipe';
+import { LengthPipe } from './pipes/length.pipe';
+import { MagicSearchPipe } from './pipes/magic-search.pipe';
+import { ManualTagSortPipe } from './pipes/manual-tags-sort.pipe';
+import { ResolutionFilterPipe } from './pipes/resolution-filter.pipe';
+import { ReturnZeroPipe } from './pipes/returnZero.pipe';
+import { SimilarityPipe } from './pipes/similarity.pipe';
+import { SortingPipe } from './pipes/sorting.pipe';
+import { StarFilterPipe } from './pipes/star-filter.pipe';
 import { TagFilterPipe } from './components/tags-auto/tag-filter.pipe';
 import { TagMatchPipe } from './components/tags-auto/tag-match.pipe';
 import { TagsDisplayPipe } from './components/tags-auto/tag-display.pipe';
-import { WordFrequencyPipe } from './components/pipes/word-frequency.pipe';
-import { WrapperPipe } from './components/pipes/wrapper.pipe';
+import { WordFrequencyPipe } from './pipes/word-frequency.pipe';
+import { WrapperPipe } from './pipes/wrapper.pipe';
 
 @NgModule({
   declarations: [
