+++ src/app/common/final-object.interface.ts
@@ -1,5 +1,5 @@
 import { ResolutionString } from '../pipes/resolution-filter.service';
-import { ScreenshotSettings } from '../../../../main-globals';
+import { ScreenshotSettings } from '../../../main-globals';
 
 export type StarRating = 0.5 | 1.5 | 2.5 | 3.5 | 4.5 | 5.5;
