+++ src/app/common/savable-properties.interface.ts
