+++ src/app/common/settings-buttons.interface.ts
