+++ src/app/common/settings-buttons.ts
