+++ src/app/common/settings-object.interface.ts
