+++ src/app/common/wizard-options.interface.ts
