+++ src/app/components/meta/meta.component.ts
@@ -3,7 +3,7 @@ import { DomSanitizer } from '@angular/platform-browser';
 
 import { ManualTagsService } from '../tags-manual/manual-tags.service';
 
-import { StarRating, ImageElement } from '../common/final-object.interface';
+import { StarRating, ImageElement } from '../../common/final-object.interface';
 import { YearEmission } from '../views/details/details.component';
 
 export interface TagEmission {
