+++ src/app/components/tags-auto/tags.component.ts
@@ -3,15 +3,15 @@ import { Component, EventEmitter, Input, Output, OnDestroy, ViewChild, ElementRe
 import { AutoTagsService, WordAndFreq } from './autotags.service';
 import { AutoTagsSaveService } from './tags-save.service';
 
-import { ImageElement } from '../common/final-object.interface';
+import { ImageElement } from '../../common/final-object.interface';
 
-import { slowFadeIn, donutAppear } from '../common/animations';
+import { slowFadeIn, donutAppear } from '../../common/animations';
 
 @Component({
   selector: 'app-tags-component',
   templateUrl: 'tags.component.html',
   styleUrls: ['../search-input.scss',
-              '../fonts/icons.scss',
+              '../../fonts/icons.scss',
               '../wizard-button.scss',
               'tags.component.scss'],
   animations: [slowFadeIn, donutAppear]
