+++ src/app/components/tags-manual/manual-tags.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 
-import { ImageElement } from '../common/final-object.interface';
+import { ImageElement } from '../../common/final-object.interface';
 
 @Injectable()
 export class ManualTagsService {
