+++ src/app/components/top/top.component.ts
@@ -4,7 +4,7 @@ import { Component, EventEmitter, Input, Output } from '@angular/core';
   selector: 'app-top-component',
   templateUrl: './top.component.html',
   styleUrls: ['./top.component.scss',
-              '../fonts/icons.scss']
+              '../../fonts/icons.scss']
 })
 export class TopComponent {
