+++ src/app/components/views/details/details.component.ts
@@ -3,7 +3,7 @@ import { DomSanitizer } from '@angular/platform-browser';
 
 import { ManualTagsService } from '../../tags-manual/manual-tags.service';
 
-import { StarRating, ImageElement } from '../../common/final-object.interface';
+import { StarRating, ImageElement } from '../../../common/final-object.interface';
 
 export interface TagEmission {
   index: number;
