+++ src/app/fonts/icons.scss
