+++ src/app/pipes/file-search.pipe.ts
