+++ src/app/pipes/folder-view.pipe.ts
