+++ src/app/pipes/magic-search.pipe.ts
