+++ src/app/pipes/manual-tags-sort.pipe.ts
@@ -1,6 +1,6 @@
 import { Pipe, PipeTransform } from '@angular/core';
 
-import { ManualTagsService } from '../tags-manual/manual-tags.service';
+import { ManualTagsService } from '../components/tags-manual/manual-tags.service';
 
 @Pipe({
   name: 'manualTagSortPipe'
