+++ src/app/pipes/resolution-filter.pipe.ts
