+++ src/app/pipes/resolution-filter.service.ts
