+++ src/app/pipes/similarity.pipe.ts
