+++ src/app/pipes/similarity.service.ts
