+++ src/app/pipes/word-frequency.pipe.ts
