+++ src/app/pipes/word-frequency.service.ts
